feat(alarm): Make AlarmNamingStrategy customizable
